% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_flows.R
\docType{data}
\name{dc_flows}
\alias{dc_flows}
\title{Washington D.C. number of inbound commuters 2020}
\format{
A data frame of 96 rows and three variables:
\describe{
\item{origin}{five digit FIPS code for county and state of origin}
\item{destination}{five digit FIPS code for county and state of destination}
\item{count}{number of workplace commuters}
}
}
\source{
https://www.irs.gov/statistics/soi-tax-stats-migration-data-2021-2022
}
\usage{
dc_flows
}
\description{
Washington D.C. number of inbound commuter counties 2020
}
\details{
From the U.S. Census Bureau:  "The American Community Survey (ACS) asks respondents about their primary workplace location. When information about workers’ residence location and workplace location are coupled, a commuting flow is generated. The origin-destination flow format informs our understanding of interconnectedness between communities, including the interchange of people, goods, and services." For additional information, see the \href{https://www.census.gov/topics/employment/commuting/guidance/flows.html}{page} on commuter data.
}
\examples{
dc_flows
}
\references{
Internal Revenue Service, Statistics of Income Division (2022). SOI Tax Stats - Migration Data. Last accessed October 26, 2024.
}
\keyword{datasets}
