% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earhart.R
\docType{data}
\name{earhart}
\alias{earhart}
\title{Global flight path and disappearance of Amelia Earhart 1937}
\format{
a data frame with 9 variables and 33 observations:
\describe{
\item{date}{date}
\item{departure_city}{departure_city}
\item{begin_lat}{begin_lat}
\item{begin_lon}{begin_lon}
\item{arrival_city}{arrival_city}
\item{end_lat}{end_lat}
\item{end_lon}{end_lon}
\item{nautical_miles}{nautical_miles}
\item{status}{status}
}
}
\source{
https://en.wikipedia.org/wiki/Amelia_Earhart
}
\usage{
earhart
}
\description{
Global flight path of Amelia Earhart 1937 and her disappearance in transit to Howland Island
}
\details{
Amelia Earhart was an American aviation pioneer and author. Among her many aviation records, she was the first female pilot to fly solo across the Atlantic in 1928, to fly solo across the North American continent and back, and the first to fly from Hawaii to Oakland, California. In 1937, she and her navigator Fred Noonan planned to circumnavigate the globe in a Lockheed Model 10-E Electra. The itinerary included 33 stops.  The trip was to begin in Oakland, California, and go eastward. After completing 30 of 33 stops, their plane went missing after they departed from Lae, New Guinea, in route to Howland Island, in the middle of the Pacific Ocean.  The two were declared lost at sea.  In 2024 an acoustic signature of an airplane was found within 100 miles of Howland Island at a depth of 16,000 ft (4.9km) of water. Some have speculated that the plane is Earhart's.
}
\examples{
earhart
}
\keyword{datasets}
