% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minard.R
\docType{data}
\name{minard_cities}
\alias{minard_cities}
\title{Cities along Napolean's march to and from Moscow 1812}
\format{
A data frame with 20 rows and three variables:
\describe{
\item{city}{city name}
\item{lon}{longitude}
\item{lat}{latitude}
}
}
\source{
https://vincentarelbundock.github.io/Rdatasets/doc/HistData/Minard.temp.html; https://www.cs.uic.edu/~wilkinson/TheGrammarOfGraphics/minard.txt
}
\usage{
minard_cities
}
\description{
Charles Joseph Minard's famous graphic of Napolean's march to Moscow contains these cities along the route.
}
\examples{
minard_cities
}
\references{
Minard, Charles Joseph. 1869. "Tableaux graphiques et cartes figuratives." Paris: Hachette.
}
\seealso{
\code{\link{minard_temps}} for the temperature data and \code{\link{minard_troops}} for the troop data.
}
\keyword{datasets}
