% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minard.R
\docType{data}
\name{minard_temps}
\alias{minard_temps}
\title{Temperature data along Napolean's march to and from Moscow 1812}
\format{
A data frame of 8 observations of four variables:
\describe{
\item{date}{date in the format of "YYYY-MM-DD"}
\item{temp}{temperature}
\item{days}{number of days on the retreat march}
\item{lon}{longitude}
}
}
\source{
https://vincentarelbundock.github.io/Rdatasets/doc/HistData/Minard.troops.html; https://www.cs.uic.edu/~wilkinson/TheGrammarOfGraphics/minard.txt
}
\usage{
minard_temps
}
\description{
Temperature played a key role in the French Army's casualties as the cold Russian winters took a toll.
}
\details{
Original documentation of this dataset directs the date to be set to the year 1815. Historical context suggests that the data is from 1812 and the year was set as such.
}
\examples{
minard_temps
}
\references{
Minard, Charles Joseph. 1869. "Tableaux graphiques et cartes figuratives." Paris: Hachette.
}
\seealso{
\code{\link{minard_cities}} for the city data and \code{\link{minard_troops}} for the troop data.
}
\keyword{datasets}
