% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_name.R
\name{state_name}
\alias{state_name}
\title{Split a string}
\usage{
state_name(name)
}
\arguments{
\item{name}{A character vector with one element.}
}
\value{
The input string split at each space.
}
\description{
Split a string
}
