% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txpyr_flows.R
\docType{data}
\name{txpyr_flows}
\alias{txpyr_flows}
\title{U.S. interstate taxpayer migration 2021 - 2022}
\format{
a data frame of 5100 observations and three variables:
\describe{
\item{dest}{two character FIPS code for the destination state}
\item{origin}{two character FIPS code for the origin state}
\item{count}{number of persons}
}
}
\source{
https://www.irs.gov/statistics/soi-tax-stats-migration-data-2021-2022
}
\usage{
txpyr_flows
}
\description{
U.S. interstate taxpayer migration data from IRS tax returns for 2021 - 2022
}
\examples{
txpyr_flows
}
\references{
Internal Revenue Service, Statistics of Income Division (2022). SOI Tax Stats - Migration Data. Last accessed October 26, 2024.
}
\keyword{datasets}
