% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txpyr_sites.R
\docType{data}
\name{txpyr_sites}
\alias{txpyr_sites}
\title{U.S. interstate taxpayer states 2021 - 2022}
\format{
a data frame of 5100 observations and three variables:
\describe{
\item{id}{two character state FIPS code}
\item{name}{two letter state postal code}
\item{lon}{longitude}
\item{lat}{latitude}
}
}
\source{
https://www.irs.gov/statistics/soi-tax-stats-migration-data-2021-2022
}
\usage{
txpyr_sites
}
\description{
U.S. taxpayer states for 2021 - 2022
}
\examples{
txpyr_sites
}
\references{
Internal Revenue Service, Statistics of Income Division (2022). SOI Tax Stats - Migration Data. Last accessed October 26, 2024.
}
\keyword{datasets}
