% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commuters.R
\docType{data}
\name{commuters}
\alias{commuters}
\title{Washington D.C. Commuters 2020}
\format{
A data frame with 97 rows and 10 columns

\describe{
\item{from_state_fips_code}{a two character FIPS code for the state of origin}
\item{from_county_fips_code}{a three character FIPS cdoe for the state of destination}
\item{from_state_name}{the full name for the state of origin}
\item{from_county_name}{a two character FIPS code for the county of origin}
\item{to_state_fips_code}{a two character FIPS code for the state of destination}
\item{to_county_fips_code}{a three character FIPS code for the county of destination}
\item{to_state_name}{the full name for the state of destination}
\item{to_county_name}{a three character FIPS code for the county of destination}
\item{workers_in_commuting_flow}{the number of workers commuting}
\item{margin_of_error}{margin of error}
}
}
\source{
https://www.census.gov/data/tables/2020/demo/metro-micro/commuting-flows-2020.html Data was created by the U.S. government and is not subject to copyright in the United States pursuant to 17 U.S.C. § 105.
}
\usage{
commuters
}
\description{
Washington D.C. inbound and outbound commuters by county 2020
}
\details{
From the U.S. Census Bureau:  "The American Community Survey (ACS) asks respondents about their primary workplace location. When information about workers’ residence location and workplace location are coupled, a commuting flow is generated. The origin-destination flow format informs our understanding of interconnectedness between communities, including the interchange of people, goods, and services." For additional information, see the \href{https://www.census.gov/topics/employment/commuting/guidance/flows.html}{page} on commuter data.
}
\examples{
commuters
}
\references{
U.S. Census Bureau, 2016-2020 5-Year ACS Commuting Flows (2020) last accessed October 24, 2024.
}
\keyword{datasets}
