% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homicides.R
\docType{data}
\name{homicides}
\alias{homicides}
\title{Homicides in San Francisco 2018 - 2023}
\format{
a dataframe with 135 rows and 7 variables:
\describe{
\item{incident_datetime}{date and time of the incident}
\item{incident_id}{unique identifier for the incident}
\item{incident_code}{incident code}
\item{incident_category}{incident category}
\item{incident_subcategory}{incident subcategory}
\item{latitude}{latitude}
\item{longitude}{longitude}
}
}
\source{
https://data.sfgov.org/Public-Safety/Police-Department-Incident-Reports-2018-to-Present/wg3w-h783/data_preview
Published under a Open Data Commons Public Domain Dedication and License v1.0 (PDDL).
}
\usage{
homicides
}
\description{
description is taken from the next paragraph. It’s shown at the top of documentation and should briefly describe the most important features of the function
}
\examples{
homicides
}
\references{
San Francisco Police Department, Police Department Incident Reports: 2018 to Present (2024) last accessed October 22, 2024.
}
\keyword{datasets}
